# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TerraScrap API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/worlds": {
      "post": {
        "summary": "Upload World",
        "operationId": "upload_world_api_worlds_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_world_api_worlds_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldCreatedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/worlds/{world_id}": {
      "get": {
        "summary": "Get World",
        "operationId": "get_world_api_worlds__world_id__get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldMetadataDto"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete World",
        "operationId": "delete_world_api_worlds__world_id__delete",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/worlds/{world_id}/tiles": {
      "get": {
        "summary": "Get Tiles",
        "operationId": "get_tiles_api_worlds__world_id__tiles_get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          },
          {
            "name": "chunk_x",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Chunk X"
            }
          },
          {
            "name": "chunk_y",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Chunk Y"
            }
          },
          {
            "name": "chunk_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 512,
              "minimum": 1,
              "default": 128,
              "title": "Chunk Size"
            }
          },
          {
            "name": "encoding",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "base64-rle-v1",
              "title": "Encoding"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TilesChunkDto"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/worlds/{world_id}/tile": {
      "get": {
        "summary": "Get Tile",
        "operationId": "get_tile_api_worlds__world_id__tile_get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          },
          {
            "name": "x",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TileDetailDto"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/worlds/{world_id}/npcs": {
      "get": {
        "summary": "Get Npcs",
        "operationId": "get_npcs_api_worlds__world_id__npcs_get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          },
          {
            "name": "town_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Town Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcListDto"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_world_api_worlds_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_world_api_worlds_post"
      },
...
      "ErrorDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorDto"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NpcDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "enum": [
              "town",
              "banner"
            ],
            "title": "Type"
          },
          "x": {
            "type": "integer",
            "title": "X"
          },
          "y": {
            "type": "integer",
            "title": "Y"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "x",
          "y"
        ],
        "title": "NpcDto"
      },
      "NpcListDto": {
        "properties": {
          "npcs": {
            "items": {
              "$ref": "#/components/schemas/NpcDto"
            },
            "type": "array",
            "title": "Npcs"
          }
        },
        "type": "object",
        "required": [
          "npcs"
        ],
        "title": "NpcListDto"
      },
...
      "WorldCreatedDto": {
        "properties": {
          "world_id": {
            "type": "string",
            "title": "World Id"
          },
          "metadata": {
            "$ref": "#/components/schemas/WorldMetadataDto"
          }
        },
        "type": "object",
        "required": [
          "world_id",
          "metadata"
        ],
        "title": "WorldCreatedDto"
      },
...
    }
  }
}